Fix RGB.decode operator-precedence bug corrupting encode round-trip
RGB.decode reassembled each channel's IEEE 754 bits with
`lr = lr + (nibble) << (4 * i)`. In Java `+` binds tighter than `<<`, so
this parsed as `(lr + nibble) << (4 * i)`, shifting the whole accumulator
every iteration instead of positioning the nibble. As a result decode did
not invert encode: a round-trip turned 0.25 into 1.727e-77. The nibble was
also an int, so once the intended shift reached 32 bits it would have
wrapped mod 32 even with correct grouping.

Cast the nibble to long and parenthesize the shift so each 4-bit group is
placed at its correct bit offset and accumulated. encode was already
correct; only decode needed the fix.

Adds RGBEncodeDecodeTest (round-trip, offset decode, multiple colors) in
engine/utils, where the other RGB tests live. Before the fix it failed
3/3; after it, RGBEncodeDecodeTest, RGBEqualsTest and ColorMathTest pass
(13 run, 0 failures).

+++ domain/color/src/main/java/org/almostrealism/color/RGB.java
@@ -730,9 +730,9 @@ public class RGB extends PackedCollection implements Externalizable, Cloneable {
 	public static RGB decode(char data[], int index) {
 		long lr = 0, lg = 0, lb = 0;
 		
-		for (int i = 0; i < 16; i++) lr = lr + (((byte)data[index + i]) - 32) << (4 * i);
-		for (int i = 0; i < 16; i++) lg = lg + (((byte)data[index + 16 + i]) - 32) << (4 * i);
-		for (int i = 0; i < 16; i++) lb = lb + (((byte)data[index + 32 + i]) - 32) << (4 * i);
+		for (int i = 0; i < 16; i++) lr = lr + ((long) (((byte) data[index + i]) - 32) << (4 * i));
+		for (int i = 0; i < 16; i++) lg = lg + ((long) (((byte) data[index + 16 + i]) - 32) << (4 * i));
+		for (int i = 0; i < 16; i++) lb = lb + ((long) (((byte) data[index + 32 + i]) - 32) << (4 * i));
 		
 		double r = Double.longBitsToDouble(lr);
 		double g = Double.longBitsToDouble(lg);
